Health-mate-api/requirements.txt
asgiref==3.11.1
Django==4.2.28
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
django-cors-headers==4.5.0
celery==5.3.6
redis==5.2.1
django-celery-beat==2.8.1
firebase-admin==6.5.0
requests==2.32.5
gunicorn==23.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
django-redis==5.4.0       
